+++ src/remote/impl/twillio.service.ts
@@ -3,15 +3,16 @@ import { Logger } from 'protocol-common/logger';
 import { ProtocolException } from 'protocol-common/protocol.exception';
 import { ProtocolErrorCode } from 'protocol-common/protocol.errorcode';
 import { Constants } from 'protocol-common/constants';
-import { SmsErrorCode } from './sms.errorcode';
+import { SmsErrorCode } from '../../sms/sms.errorcode';
 import { default as twilio } from 'twilio';
 import crypto from 'crypto';
+import { ITwillioService } from '../twillio.service.interface';
 
 /**
  * Putting Twillio in it's own service in case we want to support other SMS services
  */
 @Injectable()
-export class TwillioService {
+export class TwillioService implements ITwillioService {
 
     /**
      * send NIDP OTP SMS via Twilio
