+++ src/remote/remote.module.ts
@@ -3,6 +3,8 @@ import { IdentityService } from './impl/identity.service';
 import { AgencyService } from './impl/agency.service';
 import { IIdentityService } from './identity.service.interface';
 import { IAgencyService } from './agency.service.interface';
+import { TwillioService } from './impl/twillio.service';
+import { ITwillioService } from './twillio.service.interface';
 
 @Module({
     imports: [HttpModule],
@@ -12,7 +14,10 @@ import { IAgencyService } from './agency.service.interface';
     }, {
         provide: IAgencyService,
         useClass: AgencyService
+    }, {
+        provide: ITwillioService,
+        useClass: TwillioService
     }],
-    exports: [IIdentityService, IAgencyService]
+    exports: [IIdentityService, IAgencyService, ITwillioService]
 })
 export class RemoteModule {}
