+++ src/remote/twillio.service.interface.ts
@@ -1,0 +1,4 @@
+export abstract class ITwillioService {
+    abstract async sendOtp(toNumber: string, otp: number): Promise<void>;
+    abstract generateRandomOtp(): number;
+}
