+++ src/sms/sms.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { SmsService } from './sms.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SmsOtp } from '../entity/sms.otp';
-import { TwillioService } from './twillio.service';
 import { RateLimitModule } from '../ratelimit/ratelimit.module';
+import { RemoteModule } from '../remote/remote.module';
 
 /**
  * The SMS module doesn't expose any endpoints of it's own so doesn't need a controller
@@ -12,8 +12,9 @@ import { RateLimitModule } from '../ratelimit/ratelimit.module';
     imports: [
         TypeOrmModule.forFeature([SmsOtp]),
         RateLimitModule,
+        RemoteModule
     ],
-    providers: [SmsService, TwillioService],
+    providers: [SmsService],
     exports: [SmsService]
 })
 export class SmsModule {}
